Refactored how the file loading works and fix some issues with saving
